Refactor time utilities and improve type flexibility in chat components
