+++ client/src/components/chat/NotificationPanel.tsx
@@ -14,6 +14,7 @@ import { apiRequest } from '@/lib/queryClient';
 import { Bell, X, Check, Trash2, Users } from 'lucide-react';
 import { useNotificationManager } from '@/hooks/useNotificationManager';
 import type { ChatUser } from '@/types/chat';
+import { formatTimeAgo } from '@/utils/timeUtils';
 
 interface Notification {
   id: number;
@@ -183,16 +184,9 @@ export default function NotificationPanel({ isOpen, onClose, currentUser }: Noti
     }
   };
 
-  const formatTimeAgo = (dateString: string) => {
-    const date = new Date(dateString);
-    const now = new Date();
-    const diffInSeconds = Math.floor((now.getTime() - date.getTime()) / 1000);
-
-    if (diffInSeconds < 60) return 'الآن';
-    if (diffInSeconds < 3600) return `${Math.floor(diffInSeconds / 60)} دقيقة`;
-    if (diffInSeconds < 86400) return `${Math.floor(diffInSeconds / 3600)} ساعة`;
-    return `${Math.floor(diffInSeconds / 86400)} يوم`;
-  };
+  // استخدام الأداة العامة لتنسيق الوقت
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  const formatTimeAgoLocal = (dateString: string) => formatTimeAgo(dateString);
 
   return (
     <Dialog open={isOpen} onOpenChange={onClose}>
