+++ client/src/utils/timeUtils.ts
@@ -1,0 +1,28 @@
+export function formatTimeAgo(input: string | Date): string {
+  const date = input instanceof Date ? input : new Date(input);
+  if (isNaN(date.getTime())) return 'غير معروف';
+
+  const now = new Date();
+  const diffMs = now.getTime() - date.getTime();
+  const minutes = Math.floor(diffMs / 60000);
+
+  if (minutes < 1) return 'الآن';
+  if (minutes < 60) return `قبل ${minutes} دقيقة`;
+  const hours = Math.floor(minutes / 60);
+  if (hours < 24) return `قبل ${hours} ساعة`;
+  const days = Math.floor(hours / 24);
+  return `قبل ${days} يوم`;
+}
+
+export function getStatusColor(status?: string): string {
+  switch (status) {
+    case 'online':
+      return 'bg-green-500';
+    case 'away':
+      return 'bg-yellow-500';
+    case 'offline':
+      return 'bg-gray-400';
+    default:
+      return 'bg-gray-400';
+  }
+}
